+++ src/error.rs
@@ -166,3 +166,29 @@ impl From<Context<&str>> for TaskError {
         Self::UnexpectedError((*ctx.get_context()).into())
     }
 }
+
+/// Extension methods for `Result` types within a task body.
+///
+/// These methods can be used to convert any `Result` to the appropriate `TaskError` variant. The
+/// trait has a blanket implementation for any error type that implements
+/// [`std::error::Error`](https://doc.rust-lang.org/std/error/trait.Error.html) or
+/// [`failure::Fail`](https://docs.rs/failure/0.1.6/failure/trait.Fail.html).
+pub trait TaskResultExt<T, E> {
+    /// Convert the error type to a `TaskError::ExpectedError`.
+    fn as_expected_err(self, context: &str) -> Result<T, TaskError>;
+
+    /// Convert the error type to a `TaskError::UnexpectedError`.
+    fn as_unexpected_err(self, context: &str) -> Result<T, TaskError>;
+}
+
+impl<T, E> TaskResultExt<T, E> for Result<T, E>
+where
+    E: Fail,
+{
+    fn as_expected_err(self, context: &str) -> Result<T, TaskError> {
+        self.map_err(|_failure| TaskError::ExpectedError(context.into()))
+    }
+    fn as_unexpected_err(self, context: &str) -> Result<T, TaskError> {
+        self.map_err(|_failure| TaskError::UnexpectedError(context.into()))
+    }
+}